backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.1.0
python-multipart==0.0.20
PyPDF2==3.0.1
Pillow>=11.2.0
